+++ tests/api_connexion/endpoints/test_forward_to_fab_endpoint.py
@@ -130,7 +130,7 @@ class TestFABforwarding:
 class TestFABRoleForwarding(TestFABforwarding):
     @mock.patch("airflow.api_connexion.endpoints.forward_to_fab_endpoint.get_auth_manager")
     def test_raises_400_if_manager_is_not_fab(self, mock_get_auth_manager):
-        mock_get_auth_manager.return_value = BaseAuthManager(self.app, self.app.appbuilder)
+        mock_get_auth_manager.return_value = BaseAuthManager(self.app.appbuilder)
         response = self.client.get("api/v1/roles", environ_overrides={"REMOTE_USER": "test"})
         assert response.status_code == 400
         assert (